fix: validate benchmark comparison policies
