+++ scripts/benchmark-baseline.test.ts
@@ -6,7 +6,6 @@ import {
   benchmarkThresholdsForProfile,
   buildBenchmarkBaseline,
   comparableBenchmarksFromVitestOutput,
-  compareBenchmarkBaseline,
   defaultBenchmarkThresholds,
   grpcRuntimeBenchmarkThresholds,
   grpcRetainedRuntimeBenchmarkThresholds,
@@ -21,666 +20,41 @@ import {
   writeBenchmarkBaseline,
 } from "./benchmark-baseline.mjs";
 
-const vitestOutput = {
-  files: [
-    {
-      groups: [
-        {
-          fullName: "src/example.bench.ts > example benchmark group",
-          benchmarks: [
-            {
-              max: 3,
-              mean: 2,
-              min: 1,
-              name: "case a",
-              p99: 3,
-              sampleCount: 7,
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
-
-const summary = {
-  artifactKind: "engine-benchmark-summary",
-  backpressureCount: 0,
-  benchmarkCases: ["case a"],
-  benchmarkName: "example benchmark",
-  benchmarkScope: "engine-raw-snapshot",
-  browser: undefined,
-  cleanupLeakCount: 0,
-  groupedKeyWidthParameters: undefined,
-  groupedWriteAdmission: {
-    configuredMode: "incremental",
-    expectedAdmission: "incremental",
-  },
-  latency: {
-    outputJsonPath: "actual.json",
-    source: "vitest-output-json",
-  },
-  memory: {
-    totalDelta: {
-      rssBytes: 1024,
-    },
-  },
-  mutationCount: 100,
-  queuedEventCount: 0,
-  rowCount: 100,
-  subscriberCount: 1,
-  topics: ["orders"],
-};
-
-const runtimeHealth = {
-  engine: {
-    topics: {
-      orders: {
-        rowCount: 100,
-      },
-    },
-  },
-  kafka: {
-    topics: {
-      sourceOrders: {
-        regions: {
-          local: {
-            committedOffset: "100",
-          },
-        },
-        viewServerTopic: "orders",
-      },
-    },
-  },
-};
-
-const runtimeKafkaIngestLanes = [
-  {
-    internalTopic: "orders",
-    lane: "orders",
-    producedRows: 100,
-    region: "local",
-    sourceTopic: "sourceOrders",
-    sourceTopicAlias: "unique-topic-per-run:orders",
-  },
-];
-
-const comparableRuntimeKafkaIngestLanes = [
-  {
-    internalTopic: "orders",
-    lane: "orders",
-    producedRows: 100,
-    region: "local",
-    sourceTopicAlias: "unique-topic-per-run:orders",
-  },
-];
-
-const runtimeThroughputMutationCount = 700;
-
-const runtimeThroughputHealth = {
-  engine: {
-    topics: {
-      orders: {
-        rowCount: runtimeThroughputMutationCount,
-      },
-    },
-  },
-  kafka: {
-    topics: {
-      sourceOrders: {
-        regions: {
-          local: {
-            committedOffset: String(runtimeThroughputMutationCount),
-          },
-        },
-        viewServerTopic: "orders",
-      },
-    },
-  },
-};
-
-const runtimeThroughputKafkaIngestLanes = [
-  {
-    internalTopic: "orders",
-    lane: "orders",
-    producedRows: runtimeThroughputMutationCount,
-    region: "local",
-    sourceTopic: "sourceOrders",
-    sourceTopicAlias: "unique-topic-per-run:orders",
-  },
-];
-
-const comparableRuntimeThroughputKafkaIngestLanes = [
-  {
-    internalTopic: "orders",
-    lane: "orders",
-    producedRows: runtimeThroughputMutationCount,
-    region: "local",
-    sourceTopicAlias: "unique-topic-per-run:orders",
-  },
-];
-
-const runtimeThroughput = {
-  source: "benchmark-operation-timers",
-  cases: [
-    {
-      aggregateRowsPerSecond: 1000,
-      maxCommitObservedMs: 80,
-      maxReadSnapshotMs: 6,
-      maxTotalMs: 100,
-      meanCommitObservedMs: 75,
-      meanConvergenceMs: 75,
-      meanProducerSendMs: 25,
-      meanReadSnapshotMs: 5,
-      meanRowsPerSecond: 1000,
-      meanTotalMs: 100,
-      minRowsPerSecond: 900,
-      name: "case a",
-      producedRowsPerSample: 100,
-      readSnapshotRowsPerSample: 25,
-      sampleCount: 7,
-      totalProducedRows: 700,
-    },
-  ],
-};
-
-const comparableRuntimeThroughputCases = runtimeThroughput.cases;
-
-const rawRuntimeMetrics = {
-  eventLoopDelay: {
-    maxMs: 4,
-    meanMs: 2,
-    p99Ms: 3,
-  },
-  healthPolling: {
-    count: 11,
-    maxMs: 2,
-    totalMs: 7,
-  },
-  kafkaLag: {
-    maxConsumerLagMessages: "9007199254740993",
-    sampledRegionCount: 1,
-    totalConsumerLagMessages: "9007199254740993",
-  },
-};
-
-const runtimeMetrics = rawRuntimeMetrics;
-
-const drainedRuntimeMetrics = {
-  ...runtimeMetrics,
-  kafkaLag: {
-    maxConsumerLagMessages: "0",
-    sampledRegionCount: 1,
-    totalConsumerLagMessages: "0",
-  },
-};
-
-const comparableNonKafkaRuntimeThroughputCases = [
-  {
-    aggregateRowsPerSecond: 1000,
-    maxTotalMs: 100,
-    meanConvergenceMs: 75,
-    meanProducerSendMs: 25,
-    meanRowsPerSecond: 1000,
-    meanTotalMs: 100,
-    minRowsPerSecond: 900,
-    name: "case a",
-    producedRowsPerSample: 100,
-    sampleCount: 7,
-    totalProducedRows: 700,
-  },
-];
-
-const runtimeGrpcLeasedSample = {
-  acquiredFeedCount: 1,
-  activeLeasedFeeds: 1,
-  backpressureCount: 0,
-  cleanupActiveLeasedFeeds: 0,
-  cleanupClientActiveFeeds: 0,
-  cleanupLeakCount: 0,
-  cleanupMs: 2,
-  cleanupRowCount: 0,
-  deltaFanoutMs: 4,
-  healthOverlayMs: 1,
-  measurementRowCount: 50,
-  measuredCleanup: {
-    activeLeasedFeeds: 0,
-    activeSubscriptions: 0,
-    activeViews: 0,
-    clientActiveFeeds: 0,
-    leakCount: 0,
-    queuedEvents: 0,
-    rowCount: 0,
-  },
-  mutationCount: 50,
-  name: "gRPC leased first subscriber",
-  queuedEventCount: 0,
-  releasedFeedCount: 1,
-  rows: 50,
-  rowsPerSecond: 1000,
-  seedMutationCount: 0,
-  snapshotMs: 50,
-  subscriberCount: 1,
-  subscriptionMs: 3,
-};
-
-const runtimeGrpcLeasedOperationCaseFor = (
-  sample: typeof runtimeGrpcLeasedSample,
-  sampleCount = 7,
-) => ({
-  maxActiveLeasedFeeds: sample.activeLeasedFeeds,
-  maxCleanupActiveLeasedFeeds: sample.cleanupActiveLeasedFeeds,
-  maxCleanupClientActiveFeeds: sample.cleanupClientActiveFeeds,
-  maxCleanupMs: sample.cleanupMs,
-  maxDeltaFanoutMs: sample.deltaFanoutMs,
-  maxHealthOverlayMs: sample.healthOverlayMs,
-  maxMeasuredCleanupActiveLeasedFeeds: sample.measuredCleanup.activeLeasedFeeds,
-  maxMeasuredCleanupActiveSubscriptions: sample.measuredCleanup.activeSubscriptions,
-  maxMeasuredCleanupActiveViews: sample.measuredCleanup.activeViews,
-  maxMeasuredCleanupClientActiveFeeds: sample.measuredCleanup.clientActiveFeeds,
-  maxMeasuredCleanupLeakCount: sample.measuredCleanup.leakCount,
-  maxMeasuredCleanupQueuedEvents: sample.measuredCleanup.queuedEvents,
-  maxMeasuredCleanupRowCount: sample.measuredCleanup.rowCount,
-  maxSnapshotMs: sample.snapshotMs,
-  maxSubscriptionMs: sample.subscriptionMs,
-  meanCleanupMs: sample.cleanupMs,
-  meanDeltaFanoutMs: sample.deltaFanoutMs,
-  meanHealthOverlayMs: sample.healthOverlayMs,
-  meanRowsPerSecond: sample.rowsPerSecond,
-  meanSnapshotMs: sample.snapshotMs,
-  meanSubscriptionMs: sample.subscriptionMs,
-  medianRowsPerSecond: sample.rowsPerSecond,
-  mutationCount: sample.mutationCount * sampleCount,
-  name: sample.name,
-  pooledRowsPerSecond: sample.rowsPerSecond,
-  rowsPerSecondCoefficientOfVariation: 0,
-  sampleCount,
-  samples: Array.from({ length: sampleCount }, () => ({ ...sample })),
-  seedMutationCount: sample.seedMutationCount * sampleCount,
-});
-
-const runtimeGrpcLeasedOperationCase = runtimeGrpcLeasedOperationCaseFor(
+import {
+  vitestOutput,
+  summary,
+  runtimeHealth,
+  runtimeKafkaIngestLanes,
+  runtimeThroughputMutationCount,
+  runtimeThroughputHealth,
+  runtimeThroughputKafkaIngestLanes,
+  comparableRuntimeThroughputKafkaIngestLanes,
+  runtimeThroughput,
+  comparableRuntimeThroughputCases,
+  rawRuntimeMetrics,
+  runtimeMetrics,
+  comparableNonKafkaRuntimeThroughputCases,
   runtimeGrpcLeasedSample,
-);
-
-const runtimeGrpcLeasedReuseOperationCase = runtimeGrpcLeasedOperationCaseFor({
-  ...runtimeGrpcLeasedSample,
-  name: "gRPC leased same-route reuse",
-  subscriberCount: 10,
-});
-
-const runtimeGrpcLeasedOperationCases = [
+  runtimeGrpcLeasedOperationCaseFor,
   runtimeGrpcLeasedOperationCase,
   runtimeGrpcLeasedReuseOperationCase,
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    name: "gRPC leased one route many subscribers",
-    subscriberCount: 50,
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    name: "gRPC leased local-filter live snapshot",
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    measurementRowCount: 500,
-    mutationCount: 0,
-    name: "gRPC leased retained local-filter snapshot",
-    rows: 500,
-    seedMutationCount: 500,
-    snapshotMs: 500,
-    subscriberCount: 2,
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    measurementRowCount: 51,
-    name: "gRPC leased delta fanout",
-    seedMutationCount: 1,
-    subscriberCount: 25,
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    acquiredFeedCount: 25,
-    activeLeasedFeeds: 25,
-    measurementRowCount: 1_250,
-    mutationCount: 1_250,
-    name: "gRPC leased partitioned write convergence",
-    releasedFeedCount: 25,
-    rows: 1_250,
-    snapshotMs: 1_250,
-    subscriberCount: 25,
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    acquiredFeedCount: 25,
-    activeLeasedFeeds: 25,
-    measurementRowCount: 25,
-    mutationCount: 0,
-    name: "gRPC leased health refresh overhead",
-    releasedFeedCount: 25,
-    rows: 0,
-    rowsPerSecond: 0,
-    seedMutationCount: 25,
-    snapshotMs: 0,
-    subscriberCount: 25,
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    name: "gRPC leased last-subscriber cleanup",
-  }),
-  runtimeGrpcLeasedOperationCaseFor({
-    ...runtimeGrpcLeasedSample,
-    acquiredFeedCount: 25,
-    activeLeasedFeeds: 25,
-    measurementRowCount: 25,
-    mutationCount: 25,
-    name: "gRPC leased many routes",
-    releasedFeedCount: 25,
-    rows: 25,
-    snapshotMs: 25,
-    subscriberCount: 25,
-  }),
-];
-
-const replaceGrpcLeasedOperationCase = (
-  replacement: typeof runtimeGrpcLeasedOperationCase,
-) =>
-  runtimeGrpcLeasedOperationCases.map((operationCase) =>
-    operationCase.name === replacement.name ? replacement : operationCase,
-  );
-
-const runtimeGrpcMaterializedParameters = {
-  batchSize: 100,
-  seedRows: 1000,
-};
-
-const runtimeGrpcMaterializedSampleCount = 7;
-
-const runtimeGrpcMaterializedSampleFor = (
-  parameters: typeof runtimeGrpcMaterializedParameters,
-  name: string,
-  rows: number,
-  resultRowId: string | null,
-) => ({
-  backpressureCount: 0,
-  cleanupLeakCount: 0,
-  cleanupMs: 2,
-  healthOverlayMs: 1,
-  name,
-  queuedEventCount: 0,
-  resultRowId,
-  rows,
-  rowsPerSecond: rows * 10,
-  seedRows: parameters.seedRows,
-  snapshotMs: rows === 0 ? 0 : 7,
-  startTotalRows: parameters.seedRows,
-  streamConvergenceMs: rows === 0 ? 0 : 100,
-  totalRows: parameters.seedRows + rows,
-});
-
-const runtimeGrpcMaterializedOperationCaseFor = (
-  sample: ReturnType<typeof runtimeGrpcMaterializedSampleFor>,
-  sampleCount = runtimeGrpcMaterializedSampleCount,
-) => ({
-  maxCleanupMs: sample.cleanupMs,
-  maxHealthOverlayMs: sample.healthOverlayMs,
-  maxSnapshotMs: sample.snapshotMs,
-  maxStreamConvergenceMs: sample.streamConvergenceMs,
-  meanCleanupMs: sample.cleanupMs,
-  meanHealthOverlayMs: sample.healthOverlayMs,
-  meanRowsPerSecond: sample.rowsPerSecond,
-  meanSnapshotMs: sample.snapshotMs,
-  meanStreamConvergenceMs: sample.streamConvergenceMs,
-  medianRowsPerSecond: sample.rowsPerSecond,
-  mutationCount: sample.rows * sampleCount,
-  name: sample.name,
-  pooledRowsPerSecond: sample.rowsPerSecond,
-  rowsPerSecondCoefficientOfVariation: 0,
-  sampleCount,
-  samples: Array.from({ length: sampleCount }, () => ({ ...sample })),
-  seedMutationCount: sample.seedRows * sampleCount,
-  startTotalRows: sample.startTotalRows,
-  totalRows: sample.totalRows,
-});
-
-const runtimeGrpcMaterializedOperationCasesFor = (
-  parameters: typeof runtimeGrpcMaterializedParameters,
-) => [
-  runtimeGrpcMaterializedOperationCaseFor(
-    runtimeGrpcMaterializedSampleFor(
-      parameters,
-      "gRPC materialized stream batch",
-      parameters.batchSize,
-      `order-${parameters.seedRows + parameters.batchSize - 1}`,
-    ),
-  ),
-  runtimeGrpcMaterializedOperationCaseFor(
-    runtimeGrpcMaterializedSampleFor(
-      parameters,
-      "gRPC materialized burst",
-      parameters.batchSize * 4,
-      `order-${parameters.seedRows + parameters.batchSize * 4 - 1}`,
-    ),
-  ),
-  runtimeGrpcMaterializedOperationCaseFor(
-    runtimeGrpcMaterializedSampleFor(
-      parameters,
-      "gRPC materialized health overlay",
-      0,
-      null,
-    ),
-  ),
-];
-
-const runtimeGrpcMaterializedStreamSample = runtimeGrpcMaterializedSampleFor(
-  runtimeGrpcMaterializedParameters,
-  "gRPC materialized stream batch",
-  runtimeGrpcMaterializedParameters.batchSize,
-  "order-1099",
-);
-
-const runtimeGrpcMaterializedHealthSample = runtimeGrpcMaterializedSampleFor(
-  runtimeGrpcMaterializedParameters,
-  "gRPC materialized health overlay",
-  0,
-  null,
-);
-
-const runtimeGrpcMaterializedOperationCase = runtimeGrpcMaterializedOperationCaseFor(
-  runtimeGrpcMaterializedStreamSample,
-);
-
-const runtimeGrpcMaterializedBurstOperationCase = runtimeGrpcMaterializedOperationCaseFor(
-  runtimeGrpcMaterializedSampleFor(
-    runtimeGrpcMaterializedParameters,
-    "gRPC materialized burst",
-    runtimeGrpcMaterializedParameters.batchSize * 4,
-    "order-1399",
-  ),
-);
-
-const runtimeGrpcMaterializedHealthOperationCase = runtimeGrpcMaterializedOperationCaseFor(
-  runtimeGrpcMaterializedHealthSample,
-);
-
-const runtimeGrpcMaterializedOperationCases = [
-  runtimeGrpcMaterializedOperationCase,
-  runtimeGrpcMaterializedBurstOperationCase,
-  runtimeGrpcMaterializedHealthOperationCase,
-];
-
-const runtimeGrpcMaterializedComparisonState = (
-  operationCases: ReadonlyArray<typeof runtimeGrpcMaterializedOperationCase>,
-  caseName: string,
-) =>
-  operationCases
-    .filter((operationCase) => operationCase.name === caseName)
-    .flatMap((operationCase) =>
-      operationCase.samples.map((sample) => ({
-        backpressureCount: sample.backpressureCount,
-        cleanupLeakCount: sample.cleanupLeakCount,
-        name: sample.name,
-        queuedEventCount: sample.queuedEventCount,
-        resultRowId: sample.resultRowId,
-        rows: sample.rows,
-        seedRows: sample.seedRows,
-        startTotalRows: sample.startTotalRows,
-        totalRows: sample.totalRows,
-      })),
-    );
-
-const observation = {
-  artifactKind: "engine-benchmark-summary",
-  backpressureCount: 0,
-  benchmarks: [
-    {
-      groupName: "src/example.bench.ts > example benchmark group",
-      maxMs: 3,
-      meanMs: 2,
-      minMs: 1,
-      name: "case a",
-      p99Ms: 3,
-      sampleCount: 7,
-    },
-  ],
-  benchmarkCases: ["case a"],
-  benchmarkName: "example benchmark",
-  benchmarkScope: "engine-raw-snapshot",
-  browser: undefined,
-  cleanupLeakCount: 0,
-  groupedKeyWidthParameters: undefined,
-  groupedWriteAdmission: {
-    configuredMode: "incremental",
-    expectedAdmission: "incremental",
-  },
-  kafkaIngestLanes: undefined,
-  latencySource: "vitest-output-json",
-  memoryRssTotalDeltaBytes: 1024,
-  minimumSampleCount: 5,
-  mutationCount: 100,
-  outputJsonPath: "actual.json",
-  queuedEventCount: 0,
-  rowCount: 100,
-  seedBatchSize: undefined,
-  subscriberCount: 1,
-  summaryPath: "actual.summary.json",
-  taskLabel: "task a",
-  throughputCases: undefined,
-  topics: ["orders"],
-};
-
-const grpcLeasedObservationFor = (
-  operationCases: ReadonlyArray<typeof runtimeGrpcLeasedOperationCase>,
-) => ({
-  ...observation,
-  benchmarks: operationCases.map((operationCase) => ({
-    ...observation.benchmarks[0],
-    name: operationCase.name,
-    sampleCount: operationCase.sampleCount,
-  })),
-  benchmarkCases: operationCases.map((operationCase) => operationCase.name),
-  benchmarkScope: "runtime-grpc-leased",
-  mutationCount: operationCases.reduce(
-    (total, operationCase) => total + operationCase.mutationCount,
-    0,
-  ),
-  seedMutationCount: operationCases.reduce(
-    (total, operationCase) => total + operationCase.seedMutationCount,
-    0,
-  ),
-});
-
-const grpcLeasedObservation = grpcLeasedObservationFor([runtimeGrpcLeasedOperationCase]);
-const completeGrpcLeasedObservation = grpcLeasedObservationFor(
   runtimeGrpcLeasedOperationCases,
-);
-
-const runtimeGrpcMaterializedObservationFor = (
-  operationCases: ReadonlyArray<typeof runtimeGrpcMaterializedOperationCase>,
-) => ({
-  ...observation,
-  benchmarks: operationCases.map((operationCase) => ({
-    ...observation.benchmarks[0],
-    name: operationCase.name,
-    sampleCount: operationCase.sampleCount,
-  })),
-  benchmarkCases: operationCases.map((operationCase) => operationCase.name),
-  benchmarkScope: "runtime-grpc-materialized",
-  mutationCount: operationCases.reduce(
-    (total, operationCase) => total + operationCase.mutationCount,
-    0,
-  ),
-  seedMutationCount: operationCases.reduce(
-    (total, operationCase) => total + operationCase.seedMutationCount,
-    0,
-  ),
-});
-
-const replaceGrpcMaterializedOperationCase = (
-  replacement: typeof runtimeGrpcMaterializedOperationCase,
-) =>
-  runtimeGrpcMaterializedOperationCases.map((operationCase) =>
-    operationCase.name === replacement.name ? replacement : operationCase,
-  );
-
-const runtimeGrpcMaterializedObservation = runtimeGrpcMaterializedObservationFor(
+  runtimeGrpcMaterializedParameters,
+  runtimeGrpcMaterializedOperationCase,
   runtimeGrpcMaterializedOperationCases,
-);
+  observation,
+  grpcLeasedObservationFor,
+  grpcLeasedObservation,
+  completeGrpcLeasedObservation,
+  replaceGrpcMaterializedOperationCase,
+  runtimeGrpcMaterializedObservation,
+  runtimeGrpcMaterializedVitestOutput,
+  taskPaths,
+  browserTaskPaths,
+  runtimeTaskPaths,
+} from "./benchmark-baseline-test-fixtures.ts";
 
-const runtimeGrpcMaterializedVitestOutput = {
-  files: [
-    {
-      groups: [
-        {
-          fullName: "src/example.bench.ts > example benchmark group",
-          benchmarks: runtimeGrpcMaterializedOperationCases.map((operationCase) => ({
-            max: 3,
-            mean: 2,
-            min: 1,
-            name: operationCase.name,
-            p99: 3,
-            sampleCount: operationCase.sampleCount,
-          })),
-        },
-      ],
-    },
-  ],
-};
-
-const taskPaths = (summaryPath: string, outputJsonPath: string) => ({
-  expectedArtifactKind: "engine-benchmark-summary",
-  expectedBenchmarkScope: "engine-raw-snapshot",
-  expectedRowCount: 100,
-  label: "task a",
-  minimumSampleCount: 5,
-  outputJsonPath,
-  packageOutputJsonPath: "actual.json",
-  summaryPath,
-});
-
-const browserTaskPaths = (summaryPath: string, outputJsonPath: string) => ({
-  expectedArtifactKind: "react-browser-benchmark-summary",
-  expectedBenchmarkScope: "react-in-memory-live-query",
-  expectedRowCount: 100,
-  label: "task a",
-  minimumSampleCount: 5,
-  outputJsonPath,
-  packageOutputJsonPath: "actual.json",
-  summaryPath,
-});
-
-const runtimeTaskPaths = (summaryPath: string, outputJsonPath: string) => ({
-  expectedArtifactKind: "runtime-benchmark-summary",
-  expectedBenchmarkScope: "runtime-kafka-ingest",
-  expectedRowCount: 100,
-  label: "task a",
-  minimumSampleCount: 5,
-  outputJsonPath,
-  packageOutputJsonPath: "actual.json",
-  summaryPath,
-});
-
-describe("benchmark baseline comparison", () => {
+describe("benchmark baseline artifacts", () => {
   it("extracts comparable benchmark metrics from Vitest output", () => {
     expect(comparableBenchmarksFromVitestOutput(vitestOutput)).toStrictEqual([
       {
@@ -2212,411 +1586,6 @@ describe("benchmark baseline comparison", () => {
     );
   });
 
-  it("accepts benchmark results inside configured thresholds", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation]);
-    const actual = buildBenchmarkBaseline("smoke", [
-      {
-        ...observation,
-        benchmarks: [
-          {
-            ...observation.benchmarks[0],
-            meanMs: 15,
-            p99Ms: 23,
-          },
-        ],
-        memoryRssTotalDeltaBytes: 2048,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
-  it("reports active-query sharing structural regressions", () => {
-    const baseline = buildBenchmarkBaseline("active-query-sharing", [
-      {
-        ...observation,
-        activeViewCountBeforeCleanup: 1,
-      },
-    ]);
-    const regressed = buildBenchmarkBaseline("active-query-sharing", [
-      {
-        ...observation,
-        activeViewCountBeforeCleanup: 50,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, regressed)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: activeViewCountBeforeCleanup changed from 1 to 50.",
-      ],
-    });
-  });
-
-  it("does not force optional active-query structural counters onto older baselines", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation]);
-    const actual = buildBenchmarkBaseline("smoke", [
-      {
-        ...observation,
-        activeViewCountBeforeCleanup: 1,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
-  it("reports gRPC benchmark parameter drift", () => {
-    const changedParameters = {
-      ...runtimeGrpcMaterializedParameters,
-      batchSize: 200,
-    };
-    const changedOperationCases = runtimeGrpcMaterializedOperationCasesFor(changedParameters);
-    const baseline = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservation,
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: runtimeGrpcMaterializedOperationCases,
-      },
-    ]);
-    const actual = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservationFor(changedOperationCases),
-        grpcParameters: changedParameters,
-        runtimeOperationCases: changedOperationCases,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: mutationCount changed from 3500 to 7000.",
-        "task a: gRPC materialized stream batch runtime operation totalRows changed from 1100 to 1200.",
-        "task a: gRPC materialized stream batch runtime operation mutationCount changed from 700 to 1400.",
-        `task a: gRPC materialized stream batch runtime operation sample state changed from ${JSON.stringify(runtimeGrpcMaterializedComparisonState(runtimeGrpcMaterializedOperationCases, "gRPC materialized stream batch"))} to ${JSON.stringify(runtimeGrpcMaterializedComparisonState(changedOperationCases, "gRPC materialized stream batch"))}.`,
-        "task a: gRPC materialized burst runtime operation totalRows changed from 1400 to 1800.",
-        "task a: gRPC materialized burst runtime operation mutationCount changed from 2800 to 5600.",
-        `task a: gRPC materialized burst runtime operation sample state changed from ${JSON.stringify(runtimeGrpcMaterializedComparisonState(runtimeGrpcMaterializedOperationCases, "gRPC materialized burst"))} to ${JSON.stringify(runtimeGrpcMaterializedComparisonState(changedOperationCases, "gRPC materialized burst"))}.`,
-        'task a: grpcParameters changed from {"batchSize":100,"seedRows":1000} to {"batchSize":200,"seedRows":1000}.',
-      ],
-    });
-  });
-
-  it("reports gRPC runtime operation regressions", () => {
-    const baseline = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...completeGrpcLeasedObservation,
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: runtimeGrpcLeasedOperationCases,
-      },
-    ]);
-    const regressedOperationCase = runtimeGrpcLeasedOperationCaseFor({
-      ...runtimeGrpcLeasedSample,
-      rowsPerSecond: 50,
-      snapshotMs: 1000,
-    });
-    const regressedOperationCases = replaceGrpcLeasedOperationCase(regressedOperationCase);
-    const actual = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...grpcLeasedObservationFor(regressedOperationCases),
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: regressedOperationCases,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / gRPC leased first subscriber: meanSnapshotMs regressed from 50.000ms to 1000.000ms; allowed <= 600.000ms.",
-        "task a / gRPC leased first subscriber: maxSnapshotMs regressed from 50.000ms to 1000.000ms; allowed <= 800.000ms.",
-        "task a / gRPC leased first subscriber: pooledRowsPerSecond throughput regressed from 1000.000 rows/sec to 50.000 rows/sec; allowed >= 500.000 rows/sec.",
-      ],
-    });
-  });
-
-  it("reports materialized gRPC runtime operation regressions", () => {
-    const baseline = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservation,
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: runtimeGrpcMaterializedOperationCases,
-      },
-    ]);
-    const regressedOperationCase = {
-      ...runtimeGrpcMaterializedOperationCase,
-      maxStreamConvergenceMs: 1500,
-      meanRowsPerSecond: (6000 + 1000 / 15) / 7,
-      meanStreamConvergenceMs: 300,
-      pooledRowsPerSecond: 1000 / 3,
-      rowsPerSecondCoefficientOfVariation:
-        Math.sqrt(
-          (6 * (1000 - (6000 + 1000 / 15) / 7) ** 2 +
-            (1000 / 15 - (6000 + 1000 / 15) / 7) ** 2) /
-            7,
-        ) /
-        ((6000 + 1000 / 15) / 7),
-      samples: [
-        ...runtimeGrpcMaterializedOperationCase.samples.slice(0, 6),
-        {
-          ...runtimeGrpcMaterializedStreamSample,
-          rowsPerSecond: 1000 / 15,
-          streamConvergenceMs: 1500,
-        },
-      ],
-    };
-    const regressedOperationCases = replaceGrpcMaterializedOperationCase(
-      regressedOperationCase,
-    );
-    const actual = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservationFor(regressedOperationCases),
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: regressedOperationCases,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / gRPC materialized stream batch: pooledRowsPerSecond throughput regressed from 1000.000 rows/sec to 333.333 rows/sec; allowed >= 500.000 rows/sec.",
-      ],
-    });
-  });
-
-  it("rejects materialized gRPC runtime operation membership changes", () => {
-    const baseline = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservation,
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: runtimeGrpcMaterializedOperationCases,
-      },
-    ]);
-    const incompleteOperationCases = runtimeGrpcMaterializedOperationCases.slice(0, -1);
-    const incomplete = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservationFor(incompleteOperationCases),
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: incompleteOperationCases,
-      },
-    ]);
-
-    expect(() => compareBenchmarkBaseline(baseline, incomplete)).toThrow(
-      "Benchmark artifact field actual.tasks[0].runtimeOperationCases is missing gRPC materialized operation case: gRPC materialized health overlay.",
-    );
-
-    const unexpectedOperationCase = runtimeGrpcMaterializedOperationCaseFor(
-      runtimeGrpcMaterializedSampleFor(
-        runtimeGrpcMaterializedParameters,
-        "unexpected materialized case",
-        0,
-        null,
-      ),
-    );
-    const unexpectedOperationCases = [
-      ...runtimeGrpcMaterializedOperationCases,
-      unexpectedOperationCase,
-    ];
-    const unexpected = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservationFor(unexpectedOperationCases),
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: unexpectedOperationCases,
-      },
-    ]);
-
-    expect(() => compareBenchmarkBaseline(baseline, unexpected)).toThrow(
-      "Benchmark artifact field actual.tasks[0].runtimeOperationCases[3].name must be one of: gRPC materialized stream batch, gRPC materialized burst, gRPC materialized health overlay.",
-    );
-  });
-
-  it("rejects missing canonical leased runtime operation cases", () => {
-    const baseline = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...completeGrpcLeasedObservation,
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: runtimeGrpcLeasedOperationCases,
-      },
-    ]);
-    const incompleteOperationCases = runtimeGrpcLeasedOperationCases.slice(0, -1);
-    const actual = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...grpcLeasedObservationFor(incompleteOperationCases),
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: incompleteOperationCases,
-      },
-    ]);
-
-    expect(() => compareBenchmarkBaseline(baseline, actual)).toThrow(
-      "Benchmark artifact field actual.tasks[0].runtimeOperationCases is missing gRPC leased operation case: gRPC leased many routes.",
-    );
-  });
-
-  it("rejects unexpected leased runtime operation cases", () => {
-    const baseline = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...completeGrpcLeasedObservation,
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: runtimeGrpcLeasedOperationCases,
-      },
-    ]);
-    const unexpectedOperationCase = runtimeGrpcLeasedOperationCaseFor({
-      ...runtimeGrpcLeasedSample,
-      name: "unexpected leased case",
-    });
-    const actual = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...grpcLeasedObservationFor([
-          ...runtimeGrpcLeasedOperationCases,
-          unexpectedOperationCase,
-        ]),
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: [
-          ...runtimeGrpcLeasedOperationCases,
-          unexpectedOperationCase,
-        ],
-      },
-    ]);
-
-    expect(() => compareBenchmarkBaseline(baseline, actual)).toThrow(
-      "Benchmark artifact field actual.tasks[0].runtimeOperationCases[10].name must be one of: gRPC leased first subscriber",
-    );
-  });
-
-  it("reports runtime operation case presence changes without throwing", () => {
-    const baseline = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservation,
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: runtimeGrpcMaterializedOperationCases,
-      },
-    ]);
-    const actual = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservation,
-        benchmarkScope: "engine-raw-snapshot",
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: benchmarkScope changed from runtime-grpc-materialized to engine-raw-snapshot.",
-        "task a: seedMutationCount changed from 21000 to undefined.",
-        "task a: runtimeOperationCases presence changed.",
-        'task a: grpcParameters changed from {"batchSize":100,"seedRows":1000} to undefined.',
-      ],
-    });
-  });
-
-  it("treats zero baseline runtime operation metrics as exact invariants", () => {
-    const zeroHealthOverlayCase = {
-      ...runtimeGrpcMaterializedHealthOperationCase,
-      maxHealthOverlayMs: 0,
-      meanHealthOverlayMs: 0,
-      samples: runtimeGrpcMaterializedHealthOperationCase.samples.map((sample) => ({
-        ...sample,
-        healthOverlayMs: 0,
-      })),
-    };
-    const zeroHealthOverlayCases = replaceGrpcMaterializedOperationCase(
-      zeroHealthOverlayCase,
-    );
-    const baseline = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservationFor(zeroHealthOverlayCases),
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: zeroHealthOverlayCases,
-      },
-    ]);
-    const actual = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservation,
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: runtimeGrpcMaterializedOperationCases,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: gRPC materialized health overlay runtime operation meanHealthOverlayMs changed from 0 to 1.",
-        "task a: gRPC materialized health overlay runtime operation maxHealthOverlayMs changed from 0 to 1.",
-      ],
-    });
-  });
-
-  it("treats zero leased operation latency as an exact invariant", () => {
-    const zeroCleanupCase = {
-      ...runtimeGrpcLeasedOperationCase,
-      maxCleanupMs: 0,
-      meanCleanupMs: 0,
-      samples: runtimeGrpcLeasedOperationCase.samples.map((sample) => ({
-        ...sample,
-        cleanupMs: 0,
-      })),
-    };
-    const zeroCleanupCases = replaceGrpcLeasedOperationCase(zeroCleanupCase);
-    const baseline = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...grpcLeasedObservationFor(zeroCleanupCases),
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: zeroCleanupCases,
-      },
-    ]);
-    const actual = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...completeGrpcLeasedObservation,
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: runtimeGrpcLeasedOperationCases,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: gRPC leased first subscriber runtime operation meanCleanupMs changed from 0 to 2.",
-        "task a: gRPC leased first subscriber runtime operation maxCleanupMs changed from 0 to 2.",
-      ],
-    });
-  });
-
   it("rejects runtime operation cases that do not match benchmark cases", () => {
     expect(() =>
       validateBenchmarkBaseline({
@@ -2931,26 +1900,6 @@ describe("benchmark baseline comparison", () => {
     );
   });
 
-  it("accepts independently seeded zero-row materialized samples and compares zero throughput exactly", () => {
-    const sixSampleHealthCase = runtimeGrpcMaterializedOperationCaseFor(
-      runtimeGrpcMaterializedHealthSample,
-      6,
-    );
-    const operationCases = replaceGrpcMaterializedOperationCase(sixSampleHealthCase);
-    const baseline = buildBenchmarkBaseline("grpc-materialized", [
-      {
-        ...runtimeGrpcMaterializedObservationFor(operationCases),
-        grpcParameters: runtimeGrpcMaterializedParameters,
-        runtimeOperationCases: operationCases,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, baseline)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
   it("requires gRPC runtime operation cases in summaries and baselines", () => {
     const directory = mkdtempSync(join(tmpdir(), "view-server-benchmark-observation-"));
     const summaryPath = join(directory, "actual.summary.json");
@@ -3015,37 +1964,6 @@ describe("benchmark baseline comparison", () => {
       }),
     ).toThrow(
       "Benchmark artifact field baseline.tasks[0].runtimeOperationCases is only supported for gRPC runtime scopes.",
-    );
-  });
-
-  it("rejects gRPC runtime operation cases without operation thresholds", () => {
-    const baseline = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...completeGrpcLeasedObservation,
-        grpcParameters: {
-          retainedRows: 500,
-          routeCount: 25,
-          rowsPerFeed: 50,
-        },
-        runtimeOperationCases: runtimeGrpcLeasedOperationCases,
-      },
-    ]);
-
-    expect(() =>
-      compareBenchmarkBaseline(
-        {
-          ...baseline,
-          profile: "smoke",
-          thresholds: defaultBenchmarkThresholds,
-        },
-        {
-          ...baseline,
-          profile: "smoke",
-          thresholds: defaultBenchmarkThresholds,
-        },
-      ),
-    ).toThrow(
-      "Benchmark baseline smoke contains gRPC runtime operation cases but does not define operationMean and operationMax thresholds.",
     );
   });
 
@@ -3115,518 +2033,6 @@ describe("benchmark baseline comparison", () => {
     ).toThrow(
       "Benchmark artifact field baseline.tasks[0].grpcParameters is only supported for gRPC runtime benchmark scopes.",
     );
-  });
-
-  it("rejects large relative latency regressions for grouped order-neutral baselines", () => {
-    const subMillisecondObservation = {
-      ...observation,
-      benchmarks: [
-        {
-          ...observation.benchmarks[0],
-          meanMs: 0.2,
-          p99Ms: 0.3,
-        },
-      ],
-    };
-    const baseline = buildBenchmarkBaseline("grouped-order-neutral", [
-      subMillisecondObservation,
-    ]);
-    const actual = buildBenchmarkBaseline("grouped-order-neutral", [
-      {
-        ...subMillisecondObservation,
-        benchmarks: [
-          {
-            ...subMillisecondObservation.benchmarks[0],
-            meanMs: 1.3,
-            p99Ms: 2,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / src/example.bench.ts > example benchmark group / case a: mean regressed from 0.200ms to 1.300ms; allowed <= 1.200ms.",
-        "task a / src/example.bench.ts > example benchmark group / case a: p99 regressed from 0.300ms to 2.000ms; allowed <= 1.800ms.",
-      ],
-    });
-  });
-
-  it("keeps raw read/write sub-millisecond mean jitter report-only inside the absolute window", () => {
-    const subMillisecondObservation = {
-      ...observation,
-      benchmarks: [
-        {
-          ...observation.benchmarks[0],
-          meanMs: 0.067,
-          p99Ms: 0.3,
-        },
-      ],
-    };
-    const baseline = buildBenchmarkBaseline("raw-read-write", [subMillisecondObservation]);
-    const actual = buildBenchmarkBaseline("raw-read-write", [
-      {
-        ...subMillisecondObservation,
-        benchmarks: [
-          {
-            ...subMillisecondObservation.benchmarks[0],
-            meanMs: 0.239,
-            p99Ms: 0.3,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
-  it("rejects raw read/write mean regressions outside the focused absolute window", () => {
-    const subMillisecondObservation = {
-      ...observation,
-      benchmarks: [
-        {
-          ...observation.benchmarks[0],
-          meanMs: 0.125,
-          p99Ms: 0.3,
-        },
-      ],
-    };
-    const baseline = buildBenchmarkBaseline("raw-read-write", [subMillisecondObservation]);
-    const actual = buildBenchmarkBaseline("raw-read-write", [
-      {
-        ...subMillisecondObservation,
-        benchmarks: [
-          {
-            ...subMillisecondObservation.benchmarks[0],
-            meanMs: 0.7,
-            p99Ms: 0.3,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / src/example.bench.ts > example benchmark group / case a: mean regressed from 0.125ms to 0.700ms; allowed <= 0.625ms.",
-      ],
-    });
-  });
-
-  it("keeps raw read/write millisecond-scale runner spikes inside the focused window", () => {
-    const indexedSnapshotObservation = {
-      ...observation,
-      benchmarks: [
-        {
-          ...observation.benchmarks[0],
-          meanMs: 1.401,
-          p99Ms: 2.284,
-        },
-      ],
-    };
-    const baseline = buildBenchmarkBaseline("raw-read-write", [indexedSnapshotObservation]);
-    const actual = buildBenchmarkBaseline("raw-read-write", [
-      {
-        ...indexedSnapshotObservation,
-        benchmarks: [
-          {
-            ...indexedSnapshotObservation.benchmarks[0],
-            meanMs: 4.599,
-            p99Ms: 20.404,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
-  it("rejects raw read/write p99 regressions outside the focused tail window", () => {
-    const indexedSnapshotObservation = {
-      ...observation,
-      benchmarks: [
-        {
-          ...observation.benchmarks[0],
-          meanMs: 1.401,
-          p99Ms: 2.284,
-        },
-      ],
-    };
-    const baseline = buildBenchmarkBaseline("raw-read-write", [indexedSnapshotObservation]);
-    const actual = buildBenchmarkBaseline("raw-read-write", [
-      {
-        ...indexedSnapshotObservation,
-        benchmarks: [
-          {
-            ...indexedSnapshotObservation.benchmarks[0],
-            meanMs: 4.599,
-            p99Ms: 30,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / src/example.bench.ts > example benchmark group / case a: p99 regressed from 2.284ms to 30.000ms; allowed <= 22.284ms.",
-      ],
-    });
-  });
-
-  it("reports missing tasks, counter regressions, memory regressions, and latency regressions", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation]);
-    const regressed = buildBenchmarkBaseline("smoke", [
-      {
-        ...observation,
-        backpressureCount: 1,
-        benchmarks: [
-          {
-            groupName: "src/example.bench.ts > example benchmark group",
-            maxMs: 300,
-            meanMs: 40,
-            minMs: 1,
-            name: "case a",
-            p99Ms: 100,
-            sampleCount: 7,
-          },
-        ],
-        cleanupLeakCount: 1,
-        memoryRssTotalDeltaBytes: 200 * 1024 * 1024,
-        queuedEventCount: 1,
-      },
-    ]);
-    const withMissingTask = {
-      ...regressed,
-      tasks: [
-        ...regressed.tasks,
-        {
-          ...observation,
-          taskLabel: "extra actual task",
-        },
-      ],
-    };
-
-    expect(
-      compareBenchmarkBaseline(
-        {
-          ...baseline,
-          tasks: [
-            ...baseline.tasks,
-            {
-              ...observation,
-              taskLabel: "missing task",
-            },
-          ],
-        },
-        withMissingTask,
-      ),
-    ).toStrictEqual({
-      ok: false,
-      regressions: [
-        "extra actual task: unexpected benchmark task in actual run.",
-        "task a: cleanupLeakCount must stay 0 but was 1.",
-        "task a: backpressureCount must stay 0 but was 1.",
-        "task a: queuedEventCount must stay 0 but was 1.",
-        "task a: total RSS delta regressed from 1024 bytes to 209715200 bytes; allowed <= 3072 bytes.",
-        "task a / src/example.bench.ts > example benchmark group / case a: mean regressed from 2.000ms to 40.000ms; allowed <= 16.000ms.",
-        "task a / src/example.bench.ts > example benchmark group / case a: p99 regressed from 3.000ms to 100.000ms; allowed <= 24.000ms.",
-        "missing task: missing benchmark task in actual run.",
-      ],
-    });
-  });
-
-	  it("reports missing benchmark cases", () => {
-	    const baseline = buildBenchmarkBaseline("smoke", [observation]);
-	    const changedCases = buildBenchmarkBaseline("smoke", [
-	      {
-	        ...observation,
-	        benchmarks: [
-	          {
-	            ...observation.benchmarks[0],
-	            name: "case b",
-	          },
-	        ],
-	        benchmarkCases: ["case b"],
-	      },
-	    ]);
-
-	    expect(compareBenchmarkBaseline(baseline, changedCases)).toStrictEqual({
-	      ok: false,
-	      regressions: [
-	        'task a: benchmarkCases changed from ["case a"] to ["case b"].',
-	        "task a: unexpected benchmark case src/example.bench.ts > example benchmark group / case b.",
-	        "task a: missing benchmark case src/example.bench.ts > example benchmark group / case a.",
-	      ],
-    });
-  });
-
-  it("does not require Kafka read snapshot gates for non-Kafka throughput profiles", () => {
-    const nonKafkaThroughputObservation = {
-      ...observation,
-      throughputCases: comparableNonKafkaRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("smoke", [nonKafkaThroughputObservation]);
-    const actual = buildBenchmarkBaseline("smoke", [nonKafkaThroughputObservation]);
-
-    expect(compareBenchmarkBaseline(baseline, actual)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
-  it("requires exact Kafka ingest mutation counts", () => {
-    const kafkaObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-kafka-ingest",
-      groupedWriteAdmission: undefined,
-      kafkaIngestLanes: comparableRuntimeThroughputKafkaIngestLanes,
-      mutationCount: runtimeThroughputMutationCount,
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("kafka-ingest", [kafkaObservation]);
-    const changedMutationCount = buildBenchmarkBaseline("kafka-ingest", [
-      {
-        ...kafkaObservation,
-        mutationCount: 1400,
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            aggregateRowsPerSecond: 2000,
-            meanRowsPerSecond: 2000,
-            minRowsPerSecond: 1800,
-            producedRowsPerSample: 200,
-            totalProducedRows: 1400,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, changedMutationCount)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: mutationCount changed from 700 to 1400.",
-        "task a: case a throughput producedRowsPerSample changed from 100 to 200.",
-        "task a: case a throughput totalProducedRows changed from 700 to 1400.",
-      ],
-    });
-  });
-
-  it("requires exact WebSocket firehose mutation counts", () => {
-    const websocketObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-websocket-firehose",
-      groupedWriteAdmission: undefined,
-      mutationCount: 5,
-    };
-    const baseline = buildBenchmarkBaseline("websocket-firehose", [websocketObservation]);
-    const increasedMutationCount = buildBenchmarkBaseline("websocket-firehose", [
-      {
-        ...websocketObservation,
-        mutationCount: 10,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, increasedMutationCount)).toStrictEqual({
-      ok: false,
-      regressions: ["task a: mutationCount changed from 5 to 10."],
-    });
-  });
-
-  it("rejects task mutation counts that diverge from leased raw evidence", () => {
-    const grpcObservation = {
-      ...completeGrpcLeasedObservation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-grpc-leased",
-      grpcParameters: {
-        retainedRows: 500,
-        routeCount: 25,
-        rowsPerFeed: 50,
-      },
-      groupedWriteAdmission: undefined,
-      runtimeOperationCases: runtimeGrpcLeasedOperationCases,
-    };
-    const baseline = buildBenchmarkBaseline("grpc-leased", [grpcObservation]);
-    const increasedMutationCount = buildBenchmarkBaseline("grpc-leased", [
-      {
-        ...grpcObservation,
-        mutationCount: completeGrpcLeasedObservation.mutationCount + 1,
-      },
-    ]);
-
-    expect(() => compareBenchmarkBaseline(baseline, increasedMutationCount)).toThrow(
-      `Benchmark artifact field actual.tasks[0].runtimeOperationCases mutationCount total must equal task mutationCount ${completeGrpcLeasedObservation.mutationCount + 1} but was ${completeGrpcLeasedObservation.mutationCount}.`,
-    );
-  });
-
-  it("reports Kafka ingest throughput regressions", () => {
-    const kafkaObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-kafka-ingest",
-      groupedWriteAdmission: undefined,
-      kafkaIngestLanes: comparableRuntimeThroughputKafkaIngestLanes,
-      mutationCount: runtimeThroughputMutationCount,
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("kafka-ingest", [kafkaObservation]);
-    const regressedThroughput = buildBenchmarkBaseline("kafka-ingest", [
-      {
-        ...kafkaObservation,
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            aggregateRowsPerSecond: 400,
-            maxTotalMs: 250,
-            meanTotalMs: 250,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, regressedThroughput)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / case a: aggregateRowsPerSecond throughput regressed from 1000.000 rows/sec to 400.000 rows/sec; allowed >= 750.000 rows/sec.",
-      ],
-    });
-  });
-
-  it("reports Kafka commit-observed latency regressions", () => {
-    const kafkaObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-kafka-ingest",
-      groupedWriteAdmission: undefined,
-      kafkaIngestLanes: comparableRuntimeThroughputKafkaIngestLanes,
-      mutationCount: runtimeThroughputMutationCount,
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("kafka-ingest", [kafkaObservation]);
-    const regressedCommitObserved = buildBenchmarkBaseline("kafka-ingest", [
-      {
-        ...kafkaObservation,
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            aggregateRowsPerSecond: 100_000 / 2_100,
-            maxCommitObservedMs: 2_600,
-            maxTotalMs: 2_600,
-            meanCommitObservedMs: 2_076,
-            meanRowsPerSecond: 100_000 / 2_100,
-            meanTotalMs: 2_100,
-            minRowsPerSecond: 40,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, regressedCommitObserved)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / case a: aggregateRowsPerSecond throughput regressed from 1000.000 rows/sec to 47.619 rows/sec; allowed >= 750.000 rows/sec.",
-        "task a / case a: meanCommitObservedMs regressed from 75.000ms to 2076.000ms; allowed <= 2075.000ms.",
-        "task a / case a: maxCommitObservedMs regressed from 80.000ms to 2600.000ms; allowed <= 2580.000ms.",
-      ],
-    });
-  });
-
-  it("reports Kafka read snapshot workload and latency regressions", () => {
-    const kafkaObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-kafka-ingest",
-      groupedWriteAdmission: undefined,
-      kafkaIngestLanes: comparableRuntimeThroughputKafkaIngestLanes,
-      mutationCount: runtimeThroughputMutationCount,
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("kafka-ingest", [kafkaObservation]);
-    const regressedReadSnapshot = buildBenchmarkBaseline("kafka-ingest", [
-      {
-        ...kafkaObservation,
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            maxReadSnapshotMs: 61,
-            meanReadSnapshotMs: 41,
-            readSnapshotRowsPerSample: 10,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, regressedReadSnapshot)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: case a throughput readSnapshotRowsPerSample changed from 25 to 10.",
-        "task a / case a: meanReadSnapshotMs regressed from 5.000ms to 41.000ms; allowed <= 40.000ms.",
-        "task a / case a: maxReadSnapshotMs regressed from 6.000ms to 61.000ms; allowed <= 60.000ms.",
-      ],
-    });
-  });
-
-  it("uses the Kafka profile throughput threshold", () => {
-    const kafkaObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-kafka-ingest",
-      groupedWriteAdmission: undefined,
-      kafkaIngestLanes: comparableRuntimeThroughputKafkaIngestLanes,
-      mutationCount: runtimeThroughputMutationCount,
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("kafka-ingest", [kafkaObservation]);
-    const thresholdThroughput = buildBenchmarkBaseline("kafka-ingest", [
-      {
-        ...kafkaObservation,
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            aggregateRowsPerSecond: 750,
-            maxTotalMs: 133.33333333333334,
-            meanRowsPerSecond: 750,
-            meanTotalMs: 133.33333333333334,
-            minRowsPerSecond: 700,
-          },
-        ],
-      },
-    ]);
-    const belowThresholdThroughput = buildBenchmarkBaseline("kafka-ingest", [
-      {
-        ...kafkaObservation,
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            aggregateRowsPerSecond: 749,
-            maxTotalMs: 133.5113484646195,
-            meanRowsPerSecond: 749,
-            meanTotalMs: 133.5113484646195,
-            minRowsPerSecond: 700,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, thresholdThroughput)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-    expect(compareBenchmarkBaseline(baseline, belowThresholdThroughput)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / case a: aggregateRowsPerSecond throughput regressed from 1000.000 rows/sec to 749.000 rows/sec; allowed >= 750.000 rows/sec.",
-      ],
-    });
   });
 
   it("requires throughput cases in committed Kafka baselines", () => {
@@ -3706,37 +2112,6 @@ describe("benchmark baseline comparison", () => {
     expect(() => validateBenchmarkBaseline(baseline)).toThrow(
       "Benchmark artifact field baseline.tasks[0].throughputCases is required for runtime-kafka-sustained-firehose.",
     );
-  });
-
-  it("reports missing runtime metrics without gating noisy runtime metric values", () => {
-    const withRuntimeMetrics = {
-      ...observation,
-      runtimeMetrics,
-    };
-    const baseline = buildBenchmarkBaseline("smoke", [withRuntimeMetrics]);
-    const changedRuntimeMetrics = buildBenchmarkBaseline("smoke", [
-      {
-        ...withRuntimeMetrics,
-        runtimeMetrics: {
-          ...runtimeMetrics,
-          eventLoopDelay: {
-            maxMs: 999,
-            meanMs: 777,
-            p99Ms: 888,
-          },
-        },
-      },
-    ]);
-    const missingRuntimeMetrics = buildBenchmarkBaseline("smoke", [observation]);
-
-    expect(compareBenchmarkBaseline(baseline, changedRuntimeMetrics)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-    expect(compareBenchmarkBaseline(baseline, missingRuntimeMetrics)).toStrictEqual({
-      ok: false,
-      regressions: ["task a: runtimeMetrics presence changed."],
-    });
   });
 
   it("preserves Kafka lag precision in runtime metrics", () => {
@@ -3958,319 +2333,7 @@ describe("benchmark baseline comparison", () => {
     );
   });
 
-  it("requires drained final Kafka lag for sustained firehose baselines", () => {
-    const twoLaneKafkaIngestLanes = [
-      ...comparableRuntimeThroughputKafkaIngestLanes,
-      {
-        internalTopic: "trades",
-        lane: "trades",
-        producedRows: runtimeThroughputMutationCount,
-        region: "local",
-        sourceTopicAlias: "unique-topic-per-run:trades",
-      },
-    ];
-    const sustainedFirehoseObservation = {
-      ...observation,
-      artifactKind: "runtime-benchmark-summary",
-      benchmarkScope: "runtime-kafka-sustained-firehose",
-      groupedWriteAdmission: undefined,
-      kafkaIngestLanes: twoLaneKafkaIngestLanes,
-      mutationCount: runtimeThroughputMutationCount,
-      runtimeMetrics: {
-        ...drainedRuntimeMetrics,
-        kafkaLag: {
-          ...drainedRuntimeMetrics.kafkaLag,
-          sampledRegionCount: twoLaneKafkaIngestLanes.length,
-        },
-      },
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("kafka-sustained-firehose", [
-      sustainedFirehoseObservation,
-    ]);
-    const nonZeroLagBaseline = buildBenchmarkBaseline("kafka-sustained-firehose", [
-      {
-        ...sustainedFirehoseObservation,
-        runtimeMetrics,
-      },
-    ]);
-    const partialLagSampleBaseline = buildBenchmarkBaseline("kafka-sustained-firehose", [
-      {
-        ...sustainedFirehoseObservation,
-        runtimeMetrics: drainedRuntimeMetrics,
-      },
-    ]);
-    const missingMaxLagBaseline = buildBenchmarkBaseline("kafka-sustained-firehose", [
-      {
-        ...sustainedFirehoseObservation,
-        runtimeMetrics: {
-          ...sustainedFirehoseObservation.runtimeMetrics,
-          kafkaLag: {
-            ...sustainedFirehoseObservation.runtimeMetrics.kafkaLag,
-            maxConsumerLagMessages: null,
-          },
-        },
-      },
-    ]);
-    const missingRuntimeMetricsBaseline = buildBenchmarkBaseline("kafka-sustained-firehose", [
-      {
-        ...sustainedFirehoseObservation,
-        runtimeMetrics: undefined,
-      },
-    ]);
-    const missingKafkaIngestLanesBaseline = buildBenchmarkBaseline("kafka-sustained-firehose", [
-      {
-        ...sustainedFirehoseObservation,
-        kafkaIngestLanes: undefined,
-      },
-    ]);
-    const missingKafkaIngestLanesWithArtifactKindDriftBaseline = buildBenchmarkBaseline(
-      "kafka-sustained-firehose",
-      [
-        {
-          ...sustainedFirehoseObservation,
-          artifactKind: "engine-benchmark-summary",
-          kafkaIngestLanes: undefined,
-        },
-      ],
-    );
-
-    expect(compareBenchmarkBaseline(baseline, baseline)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-    expect(compareBenchmarkBaseline(baseline, nonZeroLagBaseline)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: runtimeMetrics.kafkaLag sampled 1 regions but expected 2.",
-        "task a: final Kafka lag must be 0 but was 9007199254740993.",
-        "task a: max final Kafka lag must be 0 but was 9007199254740993.",
-      ],
-    });
-    expect(compareBenchmarkBaseline(baseline, partialLagSampleBaseline)).toStrictEqual({
-      ok: false,
-      regressions: ["task a: runtimeMetrics.kafkaLag sampled 1 regions but expected 2."],
-    });
-    expect(compareBenchmarkBaseline(baseline, missingMaxLagBaseline)).toStrictEqual({
-      ok: false,
-      regressions: ["task a: max final Kafka lag must be 0 but was null."],
-    });
-    expect(compareBenchmarkBaseline(baseline, missingRuntimeMetricsBaseline)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: runtimeMetrics presence changed.",
-        "task a: runtimeMetrics.kafkaLag is required.",
-      ],
-    });
-    expect(() => validateBenchmarkBaseline(missingKafkaIngestLanesBaseline)).toThrow(
-      "Benchmark artifact field baseline.tasks[0].kafkaIngestLanes is required for runtime-kafka-sustained-firehose.",
-    );
-    expect(() =>
-      validateBenchmarkBaseline(missingKafkaIngestLanesWithArtifactKindDriftBaseline),
-    ).toThrow(
-      "Benchmark artifact field baseline.tasks[0].kafkaIngestLanes is required for runtime-kafka-sustained-firehose.",
-    );
-  });
-
-  it("reports throughput metadata drift", () => {
-    const withThroughput = {
-      ...observation,
-      throughputCases: comparableRuntimeThroughputCases,
-    };
-    const baseline = buildBenchmarkBaseline("smoke", [withThroughput]);
-    const missingThroughput = buildBenchmarkBaseline("smoke", [observation]);
-    const renamedThroughput = buildBenchmarkBaseline("smoke", [
-      {
-        ...withThroughput,
-        benchmarks: [
-          {
-            ...observation.benchmarks[0],
-            name: "case b",
-          },
-        ],
-        benchmarkCases: ["case b"],
-        throughputCases: [
-          {
-            ...comparableRuntimeThroughputCases[0],
-            name: "case b",
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, missingThroughput)).toStrictEqual({
-      ok: false,
-      regressions: ["task a: throughputCases presence changed."],
-    });
-    expect(compareBenchmarkBaseline(baseline, renamedThroughput)).toStrictEqual({
-      ok: false,
-      regressions: [
-        'task a: benchmarkCases changed from ["case a"] to ["case b"].',
-        "task a: unexpected throughput case case b.",
-        "task a: missing throughput case case a.",
-        "task a: unexpected benchmark case src/example.bench.ts > example benchmark group / case b.",
-        "task a: missing benchmark case src/example.bench.ts > example benchmark group / case a.",
-      ],
-    });
-  });
-
-  it("reports benchmark sample counts below the task minimum", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation]);
-    const changedSampleCount = buildBenchmarkBaseline("smoke", [
-      {
-        ...observation,
-        benchmarks: [
-          {
-            ...observation.benchmarks[0],
-            sampleCount: 1,
-          },
-        ],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, changedSampleCount)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a / src/example.bench.ts > example benchmark group / case a: sampleCount must be at least 5 but was 1.",
-      ],
-    });
-  });
-
-  it("requires exact raw write sample and mutation counts", () => {
-    const rawWriteObservation = {
-      ...observation,
-      benchmarks: [
-        {
-          ...observation.benchmarks[0],
-          sampleCount: 10,
-        },
-      ],
-      benchmarkScope: "engine-raw-write",
-      minimumSampleCount: 10,
-    };
-    const baseline = buildBenchmarkBaseline("raw-read-write", [rawWriteObservation]);
-    const changedRawWriteShape = buildBenchmarkBaseline("raw-read-write", [
-      {
-        ...rawWriteObservation,
-        benchmarks: [
-          {
-            ...rawWriteObservation.benchmarks[0],
-            sampleCount: 11,
-          },
-        ],
-        mutationCount: rawWriteObservation.mutationCount + 1,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, changedRawWriteShape)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: mutationCount changed from 100 to 101.",
-        "task a / src/example.bench.ts > example benchmark group / case a: sampleCount changed from 10 to 11.",
-      ],
-    });
-  });
-
-  it("reports benchmark metadata drift", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation]);
-    const changedMetadata = buildBenchmarkBaseline("smoke", [
-	      {
-	        ...observation,
-	        artifactKind: "react-browser-benchmark-summary",
-	        benchmarks: [
-	          {
-	            ...observation.benchmarks[0],
-	            name: "case b",
-	          },
-	        ],
-	        benchmarkCases: ["case b"],
-	        benchmarkName: "other benchmark",
-        benchmarkScope: "react-in-memory-live-query",
-        browser: {
-          browser: "firefox",
-          provider: "playwright",
-        },
-        groupedWriteAdmission: undefined,
-        groupedKeyWidthParameters: {
-          constantGroupCount: 257,
-          keyWidths: [1, 2, 4, 8],
-          orderedKeyCount: 8,
-          semanticProbe: {
-            groupByEightOrderedTotalRows: 4,
-            groupByEightTotalRows: 5,
-            groupByFourTotalRows: 3,
-            groupByOneTotalRows: 1,
-            groupByTwoTotalRows: 2,
-            orderedFirstGroupKey8: "probe-8-z",
-            orderedFirstRowCount: "10",
-            orderedSecondGroupKey8: "probe-8-y",
-            orderedSecondRowCount: "9",
-            orderedWindowRows: 4,
-          },
-          windowLimit: 250,
-        },
-        latencySource: "other-source",
-        memoryRssTotalDeltaBytes: undefined,
-        minimumSampleCount: 1,
-        mutationCount: 1,
-        outputJsonPath: "different.json",
-        rowCount: 10,
-        seedBatchSize: 10,
-        subscriberCount: 2,
-        summaryPath: "different.summary.json",
-        topics: ["trades"],
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, changedMetadata)).toStrictEqual({
-      ok: false,
-      regressions: [
-        "task a: artifactKind changed from engine-benchmark-summary to react-browser-benchmark-summary.",
-        "task a: benchmarkScope changed from engine-raw-snapshot to react-in-memory-live-query.",
-        "task a: benchmarkName changed from example benchmark to other benchmark.",
-        'task a: benchmarkCases changed from ["case a"] to ["case b"].',
-        "task a: rowCount changed from 100 to 10.",
-        "task a: mutationCount dropped from 100 to 1; allowed >= 90.",
-        "task a: subscriberCount changed from 1 to 2.",
-        'task a: topics changed from ["orders"] to ["trades"].',
-        "task a: latencySource changed from vitest-output-json to other-source.",
-        'task a: browser changed from undefined to {"browser":"firefox","provider":"playwright"}.',
-        "task a: seedBatchSize changed from undefined to 10.",
-        'task a: groupedKeyWidthParameters changed from undefined to {"constantGroupCount":257,"keyWidths":[1,2,4,8],"orderedKeyCount":8,"semanticProbe":{"groupByEightOrderedTotalRows":4,"groupByEightTotalRows":5,"groupByFourTotalRows":3,"groupByOneTotalRows":1,"groupByTwoTotalRows":2,"orderedFirstGroupKey8":"probe-8-z","orderedFirstRowCount":"10","orderedSecondGroupKey8":"probe-8-y","orderedSecondRowCount":"9","orderedWindowRows":4},"windowLimit":250}.',
-        'task a: groupedWriteAdmission changed from {"configuredMode":"incremental","expectedAdmission":"incremental"} to undefined.',
-        "task a: minimumSampleCount changed from 5 to 1.",
-	        "task a: outputJsonPath changed from actual.json to different.json.",
-	        "task a: summaryPath changed from actual.summary.json to different.summary.json.",
-	        "task a: memoryRssTotalDeltaBytes presence changed between baseline and actual run.",
-	        "task a: unexpected benchmark case src/example.bench.ts > example benchmark group / case b.",
-	        "task a: missing benchmark case src/example.bench.ts > example benchmark group / case a.",
-	      ],
-	    });
-	  });
-
-  it("accepts browser benchmark manifests without process memory when the baseline also omits memory", () => {
-    const withoutMemoryObservation = {
-      ...observation,
-      artifactKind: "react-browser-benchmark-summary",
-      benchmarkScope: "react-in-memory-live-query",
-      memoryRssTotalDeltaBytes: undefined,
-    };
-    const baseline = buildBenchmarkBaseline("smoke", [withoutMemoryObservation]);
-    const withoutMemory = buildBenchmarkBaseline("smoke", [
-      {
-        ...withoutMemoryObservation,
-        memoryRssTotalDeltaBytes: undefined,
-      },
-    ]);
-
-    expect(compareBenchmarkBaseline(baseline, withoutMemory)).toStrictEqual({
-      ok: true,
-      regressions: [],
-    });
-  });
-
   it("rejects missing actual RSS data for engine baselines", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation]);
     const withoutMemory = buildBenchmarkBaseline("smoke", [
       {
         ...observation,
@@ -4278,7 +2341,7 @@ describe("benchmark baseline comparison", () => {
       },
     ]);
 
-    expect(() => compareBenchmarkBaseline(baseline, withoutMemory)).toThrow(
+    expect(() => validateBenchmarkBaseline(withoutMemory, "actual")).toThrow(
       "Benchmark artifact field actual.tasks[0].memoryRssTotalDeltaBytes is required for engine-benchmark-summary.",
     );
   });
@@ -4456,29 +2519,9 @@ describe("benchmark baseline comparison", () => {
   it("rejects empty baseline task manifests", () => {
     const baseline = buildBenchmarkBaseline("smoke", []);
 
-    expect(() => compareBenchmarkBaseline(baseline, baseline)).toThrow(
+    expect(() => validateBenchmarkBaseline(baseline)).toThrow(
       "Benchmark artifact field baseline.tasks must be a non-empty array.",
     );
   });
 
-  it("rejects duplicate task labels in baseline manifests", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [observation, observation]);
-
-    expect(() => compareBenchmarkBaseline(baseline, baseline)).toThrow(
-      "Benchmark artifact field baseline.tasks contains duplicate taskLabel: task a.",
-    );
-  });
-
-  it("rejects duplicate benchmark cases in baseline manifests", () => {
-    const baseline = buildBenchmarkBaseline("smoke", [
-      {
-        ...observation,
-        benchmarks: [observation.benchmarks[0], observation.benchmarks[0]],
-      },
-    ]);
-
-    expect(() => compareBenchmarkBaseline(baseline, baseline)).toThrow(
-      "Benchmark artifact field baseline.tasks[task a].benchmarks contains duplicate benchmark case: src/example.bench.ts > example benchmark group / case a.",
-    );
-  });
 });
