+++ vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
         "scripts/benchmark-artifact-mechanics.mjs",
         "scripts/benchmark-artifact-io.mjs",
         "scripts/benchmark-baseline.mjs",
+        "scripts/benchmark-comparison-policy.mjs",
         "scripts/benchmark-baseline-cli.mjs",
         "scripts/benchmark-baseline-profiles.mjs",
         "scripts/benchmark-baseline-workflow.mjs",
